Authorization

+++ Restaurants.Application/Dishes/Commands/DeleteDish/DeleteDishByIdForRestaurantCommandHandler.cs
@@ -2,15 +2,18 @@
 using Microsoft.Extensions.Logging;
 using Restaurants.Application.Dishes.Commands.DeleteDishes;
 using Restaurants.Application.Dishes.Queries.GetDishByIdForRestaurant;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDish
 {
     public class DeleteDishByIdForRestaurantCommandHandler(ILogger<DeleteDishesForRestaurantCommandHandler> _logger,
         IRestaurantsRepository _restaurantsRepository,
-        IDishesRepository _dishesRepository) : IRequestHandler<DeleteDishByIdForRestaurantCommand>
+        IDishesRepository _dishesRepository,
+        IRestaurantAuthoirzationService _restaurantAuthoirzationService) : IRequestHandler<DeleteDishByIdForRestaurantCommand>
     {
         public async Task Handle(DeleteDishByIdForRestaurantCommand request, CancellationToken cancellationToken)
         {
@@ -22,11 +25,16 @@ namespace Restaurants.Application.Dishes.Commands.DeleteDish
                 throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
             }
 
-            var dish = restaurant.Dishes.FirstOrDefault(d=>d.DishId==request.DishId);
+            var dish = await _dishesRepository.GetDishByIdForRestaurant(request.RestaurantId,request.DishId);
 
             if (dish is null)
             {
                 throw new NotFoundException(nameof(Dish), request.DishId.ToString());
+            }
+
+            if (!_restaurantAuthoirzationService.Authorize(restaurant, ResourceOperation.Update))
+            {
+                throw new ForbidException();
             }
 
             await _dishesRepository.Delete(dish);
